split the url status into new test file
