+++ scout/test/navigation.py
@@ -148,25 +148,6 @@ class NavigationTest(LiveServerTestCase):
         self.assertLocation('/')
 
     # @wd.parallel.multiply
-    def test_home_exists(self):
-        """Test that the home page results in a 200"""
-        self.updateSauceName('Pageflow: Home Filter 200')
-        # Home Page
-        self.assertUrlStatus('/', 200)
-
-    # @wd.parallel.multiply
-    def test_food_exists(self):
-        """Test that the food page results in a 200"""
-        self.updateSauceName('Pageflow: Food Filter 200')
-        self.assertUrlStatus('/food/', 200)
-
-    # @wd.parallel.multiply
-    def test_filter_exists(self):
-        """Test that the filter page results in a 200"""
-        self.updateSauceName('Pageflow: Filter URL 200')
-        self.assertUrlStatus('/filter/', 200)
-
-    # @wd.parallel.multiply
     def test_home_content(self):
         """Test that there is at least one place listed on the home page"""
         self.updateSauceName('Pageflow: Home Content')
@@ -225,28 +206,3 @@ class NavigationTest(LiveServerTestCase):
         self.go_url('/')
         clickable = self.driver.find_element_by_id('link_discover')
         self.assertEqual(clickable.get_attribute('disabled'), 'true')
-
-    # @wd.parallel.multiply
-    def test_bad_detailURL(self):
-        """Ensure a nonexistant space results in a 404 status code"""
-        self.updateSauceName('Pageflow: Nonexistant Space Details Page')
-        self.assertUrlStatus('/detail/12345679', 404)
-
-    # uncomment for debugging
-    # @unittest.expectedFailure
-    # @wd.parallel.multiply
-    def test_bad_homeURL(self):
-        """Test an invalid URL and see if it results in a 404"""
-        self.updateSauceName('Pageflow: Bad Home URL')
-        # Bad URL
-        self.assertUrlStatus('/LSFDLK/', 404)
-
-    # uncomment for debugging
-    # @unittest.expectedFailure
-    # @wd.parallel.multiply
-    def test_redirect_URLs(self):
-        """Test URLs that are meant to redirect(301)"""
-        self.updateSauceName('Pageflow: Redirecting URLs')
-        # Test some other potential url's that should be able to redirect
-        self.assertUrlStatus('/food', 301)
-        self.assertUrlStatus('/filter', 301)
